Refactor time series tooltip
